move config to src/ dir
