+++ src/actions/sockets.js
@@ -2,7 +2,7 @@
 import SocketIOClient from 'socket.io-client';
 import * as types from '../constants/sockets';
 import { redirect } from './services';
-import config from '../../config';
+import config from '../config';
 
 export const missingSocketConnection = () => ({
   type: types.SOCKETS_CONNECTION_MISSING,
